+++ app/src/main/java/ort/tp3_login/holders/ServicioViewHolder.kt
@@ -29,7 +29,7 @@ class ServicioViewHolder(v: View) : RecyclerView.ViewHolder(v) {
         this.view = v
     }
 
-    fun setNames (titulo: String, nombreGuia: String, imageView: Uri, valoracion: Int, perfilpic: Int, activityId: String,user:UsuarioLogin, token:String) {
+    fun setNames (titulo: String, nombreGuia: String, imageView: Uri, valoracion: Double, perfilpic: Int, activityId: String,user:UsuarioLogin, token:String) {
         var textViewTitulo : TextView = view.findViewById(R.id.tituloCard)
         var textViewNombreGuia : TextView = view.findViewById(R.id.nombreGuiaCard)
         var imagenView : ImageView = view.findViewById(R.id.imageCard)
